+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - refactor(core): make the power-off menu, a sub-menu with power-off and reboot action items - closes #123
 - fix(core): headed menus not showing the first item in the list - closes #144
 - refactor(system): generate the hostname of the device based on a hash of its serial number
+- feat(core): show hostname of device on the home page - closes #128
 
 ## Version 0.15.4